removed hard coded nersc paths, making them environment variables removed nersc-specific variable names and comments copied field_extents.fits to scratch
