+++ REQUIRE
@@ -1,5 +1,4 @@
 julia 0.5-
-Compat
 DataFrames
 Distributions
 DocOpt
