+++ src/Celeste.jl
@@ -14,8 +14,8 @@ include("Infer.jl")
 
 # public API
 export infer, score_field
+include("stage.jl")
 include("api.jl")
-include("nersc_io.jl")
 include("score.jl")
 
 end # module
